added user profile picture upload
